add some feature
