+++ models/cart.js
@@ -1,74 +1,70 @@
 'use strict';
-const { hashPassword } = require("../helpers/bcrypt")
 const {
   Model
 } = require('sequelize');
 module.exports = (sequelize, DataTypes) => {
-  class User extends Model {
+  class Cart extends Model {
     /**
      * Helper method for defining associations.
      * This method is not a part of Sequelize lifecycle.
      * The `models/index` file will call this method automatically.
      */
     static associate(models) {
-      // define association here
+      Cart.belongsTo(models.User)
+      Cart.belongsTo(models.Product)
     }
   };
-  User.init({
-    email: {
-      type: DataTypes.STRING,
+  Cart.init({
+    UserId: {
+      type: DataTypes.INTEGER,
       allowNull: false,
       validate: {
-        isEmail: {
-          args: true,
-          msg: 'Invalid email format'
-        },
         notNull: {
           args: true,
-          msg: 'email is required'
+          msg: 'UserId required'
         },
         notEmpty: {
           args: true,
-          msg: 'email is required'
+          msg: 'UserId required'
         }
       }
     },
-    password: {
-      type: DataTypes.STRING,
+    ProductId: {
+      type: DataTypes.INTEGER,
       allowNull: false,
       validate: {
         notNull: {
           args: true,
-          msg: 'password is required'
+          msg: 'ProductId required'
         },
         notEmpty: {
           args: true,
-          msg: 'password is required'
+          msg: 'ProductId required'
         }
       }
     },
-    role: {
-      type: DataTypes.STRING,
+    quantity: {
+      type: DataTypes.INTEGER,
       allowNull: false,
       validate: {
         notNull: {
           args: true,
-          msg: 'role is required'
+          msg: 'quantity required'
         },
         notEmpty: {
           args: true,
-          msg: 'role is required'
+          msg: 'quantity required'
+        },
+        min: {
+          args: [1],
+          msg: 'quantity must be more than 0'
         }
       }
-    }
+    },
+    status: DataTypes.BOOLEAN
   }, {
     sequelize,
-    modelName: 'User',
-    hooks: {
-      beforeCreate (user) {
-        user.password = hashPassword(user.password)
-      }
-    }
+    modelName: 'Cart',
   });
-  return User;
+  return Cart;
 };
