+++ models/user.js
@@ -1,0 +1,65 @@
+'use strict';
+const { hashPassword } = require("../helpers/bcrypt")
+const {
+  Model
+} = require('sequelize');
+module.exports = (sequelize, DataTypes) => {
+  class User extends Model {
+    /**
+     * Helper method for defining associations.
+     * This method is not a part of Sequelize lifecycle.
+     * The `models/index` file will call this method automatically.
+     */
+    static associate(models) {
+      User.hasMany(models.Cart)
+    }
+  };
+  User.init({
+    email: {
+      type: DataTypes.STRING,
+      allowNull: false,
+      validate: {
+        isEmail: {
+          args: true,
+          msg: 'Invalid email format'
+        },
+        notNull: {
+          args: true,
+          msg: 'email is required'
+        },
+        notEmpty: {
+          args: true,
+          msg: 'email is required'
+        }
+      }
+    },
+    password: {
+      type: DataTypes.STRING,
+      allowNull: false,
+      validate: {
+        notNull: {
+          args: true,
+          msg: 'password is required'
+        },
+        notEmpty: {
+          args: true,
+          msg: 'password is required'
+        }
+      }
+    },
+    role: {
+      type: DataTypes.STRING,
+      allowNull: false
+    }
+  }, {
+    sequelize,
+    modelName: 'User',
+    hooks: {
+      beforeCreate (user) {
+        user.password = hashPassword(user.password)
+        user.role = 'Customer'
+      }
+    }
+  });
+  return User;
+};
