--- server/bin/http
@@ -1,4 +1,0 @@
-const app = require('../app.js')
-const PORT = process.env.PORT || 3000
-
-app.listen(PORT, () => console.log(`App listen on port ${PORT}`))
