--- server/package.json
@@ -1,26 +1,0 @@
-{
-  "name": "server",
-  "version": "1.0.0",
-  "description": "",
-  "main": "index.js",
-  "scripts": {
-    "test": "jest --detectOpenHandles --forceExit --runInBand --verbose",
-    "dev": "nodemon bin/http",
-    "start": "node bin/http"
-  },
-  "keywords": [],
-  "author": "",
-  "license": "ISC",
-  "devDependencies": {
-    "jest": "^26.6.3",
-    "supertest": "^6.0.1"
-  },
-  "dependencies": {
-    "bcryptjs": "^2.4.3",
-    "cors": "^2.8.5",
-    "express": "^4.17.1",
-    "jsonwebtoken": "^8.5.1",
-    "pg": "^8.4.2",
-    "sequelize": "^6.3.5"
-  }
-}
