--- server/seeders/20201109055643-seeding-Users-table.js
@@ -1,19 +1,0 @@
-'use strict';
-
-const { hashPassword } = require("../helpers/bcrypt")
-
-module.exports = {
-  up: async (queryInterface, Sequelize) => {
-    await queryInterface.bulkInsert('Users', [{
-      email: 'admin@mail.com',
-      password: hashPassword('1234'),
-      role: 'admin',
-      createdAt: new Date(),
-      updatedAt: new Date()
-    }])
-  },
-
-  down: async (queryInterface, Sequelize) => {
-    await queryInterface.bulkDelete('Users')
-  }
-};
